requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
pyarrow>=12.0.0
python-dotenv>=1.0.0

